+++ NEWS.md
@@ -1,6 +1,10 @@
 # csm (development version)
 * Added basic functions for data quality check
-* New function to select a center from the output given by `nest_tables`
+* `select_center`: function to select a center from the output given by
+  `nest_tables`.
+* `render_report`: function that automatically renders reports for 
+  more than one center involved in the study.
+* Added `rmarkdown` as imports.
 * Fix `.travis.yml`
 
 # csm 0.0.0.9000